Rename to \VuFind\Form\Form
