+++ module/VuFind/src/VuFind/Form/Form.php
@@ -25,7 +25,7 @@
  * @license  http://opensource.org/licenses/gpl-2.0.php GNU General Public License
  * @link     https://vufind.org/wiki/development:plugins:controllers Wiki
  */
-namespace VuFind;
+namespace VuFind\Form;
 
 use VuFind\Config\YamlReader;
 use Zend\InputFilter\InputFilter;
